spring

+++ SpMVC_ReadBook/src/main/java/com/biz/book/mapper/BookDao.java
@@ -2,4 +2,6 @@ package com.biz.book.mapper;
 
 import com.biz.book.model.BookVO;
 
-public interface BookDao extends GenericDao<BookVO,Integer> {}
+public interface BookDao extends GenericDao<BookVO, Long> {
+	
+}
